CREATCMR-7137: GCARS Source name sequence

+++ CreateCMRBatch/src/com/ibm/cmr/create/batch/service/GCARSService.java
@@ -174,17 +174,17 @@ public class GCARSService extends MultiThreadedBatchService<GCARSUpdtQueue> {
       GCARSUpdtQueue queue = query.getSingleResult(GCARSUpdtQueue.class);
       String lastSeq = null;
       String lastSourceName = null;
-      int nextSeq = 0;
+      Long nextSeq = 0L;
 
       if (queue != null) {
         lastSourceName = queue.getId().getSourceName();
         lastSeq = lastSourceName.substring((lastSourceName.indexOf("-")) + 1);
       }
       if (!StringUtils.isBlank(lastSeq)) {
-        nextSeq = Integer.valueOf(lastSeq) + 1;
+        nextSeq = Long.valueOf(lastSeq) + 1;
         LOG.debug("GCARS next sequence is " + nextSeq);
       } else {
-        nextSeq = 1;
+        nextSeq = 1L;
       }
       int lineNumber = 1;
